[FEAT] 검색 결과 보여주기 (#119)
